cleaning scenes
Former-commit-id: 543e1efb0b7bb7f13395bfc481af1dde3d3846af

+++ applications/plugins/SofaTest/BroadPhase_test.h
@@ -159,7 +159,7 @@ void randMoving(sofa::core::CollisionModel* cm,const sofa::defaulttype::Vector3 
     sofa::core::objectmodel::Data<MechanicalObjectRigid3::VecDeriv> & dvelocities = *dof->write( sofa::core::VecId::velocity() );
     MechanicalObjectRigid3::VecDeriv & velocities = *dvelocities.beginEdit();
 
-    for(int i = 0 ; i < dof->getSize() ; ++i){
+    for(size_t i = 0 ; i < dof->getSize() ; ++i){
         if( (sofa::helper::irand()) < RAND_MAX/2.0){//make it move !
             velocities[i] = sofa::defaulttype::Vector3(1,1,1);//velocity is used only to know if a primitive moves, its direction is not important
             positions[i] = sofa::defaulttype::Rigid3Types::Coord(randVect(min_vect,max_vect),sofa::defaulttype::Quaternion(0,0,0,1));
